reimplement CrossMappingCardinality function to compute CMC causal strength and cross mapping AUC scores

+++ src/GCCM4Lattice.cpp
@@ -194,13 +194,13 @@ std::vector<std::vector<double>> GCCM4Lattice(
   std::vector<int> unique_lib_sizes(lib_sizes.begin(), lib_sizes.end());
 
   // Transform to ensure no size exceeds max library size
-  int max_lib_size = lib.size();
+  int max_lib_size = static_cast<int>(lib.size());
   std::transform(unique_lib_sizes.begin(), unique_lib_sizes.end(), unique_lib_sizes.begin(),
                  [&](int size) { return std::min(size, max_lib_size); });
 
-  // Ensure the minimum value in unique_lib_sizes is E + 2 (uncomment this section if required)
+  // Ensure the minimum value in unique_lib_sizes is b (uncomment this section if required)
   // std::transform(unique_lib_sizes.begin(), unique_lib_sizes.end(), unique_lib_sizes.begin(),
-  //                [&](int size) { return std::max(size, E + 2); });
+  //                [&](int size) { return std::max(size, b); });
 
   // Remove duplicates
   std::sort(unique_lib_sizes.begin(), unique_lib_sizes.end());
